refactor(meeting-tools): move Zoom inline tools to skills/zoom/tools.ts
Satisfies the CLAUDE.md rule that inline tools live only in
src/inline-tools.ts or src/browser-tools.ts (issue #786).

- New: skills/zoom/tools.ts — summonTool, dismissTool, joinZoomTool
  (plus their Zoom-specific helpers getZoomPMI/Passcode/ShareScreen)
- src/meeting-tools.ts: stripped to joinGmeetTool, lookupMeetingIdTool,
  callContactTool; removed execFileSync import no longer needed
- src/inline-tools.ts: re-exports Zoom trio from ../skills/zoom/tools.js;
  remaining three still from ./meeting-tools.js; no change to voice-agent.ts
